Rename 'method' module to 'methods' (plural)
