+++ crates/dids/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod bearer;
 pub mod document;
 pub mod identifier;
-pub mod method;
+pub mod methods;
 pub mod resolver;
