+++ crates/dids/src/methods/dht/bep44.rs
