+++ crates/dids/src/methods/spruce_mappers/document.rs
