+++ crates/dids/src/methods/spruce_mappers/document_metadata.rs
