+++ crates/dids/src/methods/spruce_mappers/resolution_result.rs
