+++ crates/dids/src/methods/web.rs
@@ -1,4 +1,4 @@
-use crate::method::{Method, ResolutionResult, Resolve};
+use crate::methods::{Method, ResolutionResult, Resolve};
 use did_web::DIDWeb as SpruceDidWebMethod;
 use ssi_dids::did_resolve::{DIDResolver, ResolutionInputMetadata};
